Handle cv and cm as proper operators
An operator normaly takes a motion: `d {motion}`
But if you use an operator twice it will act linweise: `dd`
Which is equivalent to `d_` but its more comfortable to just repeat the
operator key

We have two mappings acting as operator for the extract variable and
method operations.
Respectively `cv` and `cm`
So we would expect `cvv` and `cmm` to work linewise, this is what this
commit does.